refactor: profile 쿼리 위치이동
