+++ src/components/myPage/Profile.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useProfilesQuery } from '@/hooks/queries/byUse/useProfilesQueries';
+import { useProfilesQuery } from '@/hooks/queries/profiles/useProfilesQueries';
 import { Button } from '@/stories/Button';
 import Spinner from '@/stories/Spinner';
 import Link from 'next/link';
@@ -12,22 +12,22 @@ const Profile = ({ userId }: { userId: string }) => {
   if (isProfileError) return <>오류...</>;
   if (isProfileLoading)
     return (
-      <div className='w-full h-full absolute z-[3000]  flex justify-center items-center'>
+      <div className='absolute z-[3000] flex h-full w-full items-center justify-center'>
         <Spinner />
       </div>
     );
 
   return (
-    <div className='flex flex-col items-center mt-12 mb-16'>
-      <div className='w-[184px] h-[184px] overflow-hidden rounded-full'>
+    <div className='mb-16 mt-12 flex flex-col items-center'>
+      <div className='h-[184px] w-[184px] overflow-hidden rounded-full'>
         <img
           alt='프로필 이미지'
           src={profileData?.profileImageUrl || '/svgs/Profile.svg'}
-          className='object-cover w-full h-full'
+          className='h-full w-full object-cover'
         />
       </div>
-      <span className='mt-6 text-black text-title_xl'>{profileData?.profiles.user_nickname || '닉네임 없음'}</span>
-      <span className='mt-2 text-black text-body_lg'>{profileData?.profiles.user_email}</span>
+      <span className='mt-6 text-title_xl text-black'>{profileData?.profiles.user_nickname || '닉네임 없음'}</span>
+      <span className='mt-2 text-body_lg text-black'>{profileData?.profiles.user_email}</span>
       <div className='mt-8'>
         <Link href={'/mypage/edit'}>
           <Button
@@ -39,7 +39,7 @@ const Profile = ({ userId }: { userId: string }) => {
             <img
               alt='프로필 수정'
               src={'/svgs/Setting.svg'}
-              className='object-cover w-full h-full'
+              className='h-full w-full object-cover'
             />
           </Button>
         </Link>
