+++ src/components/myPage/ProfileUpdate.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import { useUpdateProfile } from '@/hooks/queries/byUse/useProfileMutation';
-import { useProfilesQuery } from '@/hooks/queries/byUse/useProfilesQueries';
 import { useUploadImage } from '@/hooks/queries/byUse/useStorageMutation';
+import { useUpdateProfile } from '@/hooks/queries/profiles/useProfileMutation';
+import { useProfilesQuery } from '@/hooks/queries/profiles/useProfilesQueries';
 import { useProfileForm } from '@/hooks/useCustomForm/useProfileForm';
 import { Button } from '@/stories/Button';
 import { useRouter } from 'next/navigation';
